Test higher timeouts to load test gcloud function
We're trying to understand gcloud functions behavior when set as a backend to Fastly. With this setting in place should just pass requests to gcloud function and prove that gcloud function will queue up requests when it auto scales. Without these settings, probe to gcloud function returns first_byte_timeout error since it is auto-scaling.