Mwcm 2370/slash marketing and mwcm redirect updates (#585)
* Update to send all of /marketing to MWCM across all environments. Now /subscriptions is the only path remaining with the dev condition.

Added whitelist query string parameter for MWCM requests within /subscriptions to send the ptr arg to the backend. This is needed because we have to configure some redirects that vary by this arg.

Also fixed how query string args are handled on redirects for when both the original request and redirect location include query string args. Previously this would result in something like ?originalRequestParam1=foo&originalRequestParam2=bar?redirectLocationParam=baz which is invalid.

* Whitelisted ptr query string arg for /subscriptions

* fix query string ampersand

* add tests

* add stage flag for  /marketing

* update tests for add stage flag for  /marketing

* unset staging flag from previous commit

* fix whitespace