Add the Video Offsite Player (#840)
This is GDPR-related: the video offsite player is currently being served
by static01.nyt.com, but that really doesn't make sense as the offsite
player is an application, not a static asset. The player will need to
have access to the GDPR and tracking cookies, and those are going to be
exposed only on nytimes.com.

We could migrate it to static01.nytimes.com instead, but that would
require replicating the gdpr.vcl logic on that service.